refactor(server): decompose browser toolset into browser/ subdirectory
Split tools/toolsets/browser.ts (982 lines) into focused modules:
- browser/schemas.ts — all 8 Zod input schemas and shared field definitions
- browser/operations.ts — executeOperation, shouldReturnFreshSnapshot, actionTerminatesSequence
- browser/formatters.ts — output formatters and withFreshSnapshot
- browser/queue.ts — runSerialized serialised-execution harness and runBrowserTool
- browser/index.ts — thin wiring, createBatchTool, createBrowserToolset

Also moved browser-prune-policy and browser-snapshot-serializer into the
subdirectory, renamed to prune-policy and snapshot-serializer. Updated all
import paths accordingly.
