+++ packages/server/src/tools/toolsets/browser/snapshot-serializer.ts
