+++ packages/server/src/tools/toolsets/register-default-toolsets.ts
@@ -1,5 +1,5 @@
 import { createAgendaToolset } from '@/tools/toolsets/agenda.js';
-import { createBrowserToolset } from '@/tools/toolsets/browser.js';
+import { createBrowserToolset } from '@/tools/toolsets/browser/index.js';
 import { createRecordingsToolset } from '@/tools/toolsets/recordings.js';
 import { registerToolset } from '@/tools/toolsets/registry.js';
 import { createSessionHistoryToolset } from '@/tools/toolsets/session-history.js';
